add changelog

+++ src/main/java/org/opensearch/knn/common/KNNVectorUtil.java
@@ -16,7 +16,7 @@ public class KNNVectorUtil {
      */
     public static boolean isZeroVector(byte[] vector) {
         for (byte e : vector) {
-            if (e != (byte) 0) {
+            if (e != 0) {
                 return false;
             }
         }
